+++ src/main/java/com/paulo/minipicpay/model/UserRepository.java
@@ -1,8 +1,10 @@
 package com.paulo.minipicpay.model;
 
+import java.math.BigDecimal;
+
 public interface UserRepository {
 
     void insert(User user);
-    void updateBalance(User user, Double amount);
+    void updateBalance(User user, BigDecimal amount);
     User findBy(Document document);
 }
