+++ src/main/java/com/paulo/minipicpay/model/ValidateTransferHandler.java
@@ -1,6 +1,8 @@
 package com.paulo.minipicpay.model;
 
+import java.math.BigDecimal;
+
 public interface ValidateTransferHandler {
 
-    void validate(User user, Double amount);
+    void validate(User user, BigDecimal amount);
 }
